OPHTUTU-417: hakemusSearchRepository + imports
