+++ tutu-backend/src/main/scala/fi/oph/tutu/backend/repository/MaakoodiRepository.scala
@@ -9,7 +9,6 @@ import slick.jdbc.GetResult
 import slick.jdbc.PostgresProfile.api.*
 
 import java.util.UUID
-import scala.concurrent.duration.DurationInt
 import scala.util.{Failure, Success}
 import scala.concurrent.ExecutionContext.Implicits.global
 
@@ -19,8 +18,7 @@ class MaakoodiRepository {
   @Autowired
   val db: TutuDatabase = null
 
-  final val DB_TIMEOUT = 30.seconds
-  val LOG: Logger      = LoggerFactory.getLogger(classOf[MaakoodiRepository])
+  val LOG: Logger = LoggerFactory.getLogger(classOf[MaakoodiRepository])
 
   implicit val getMaakoodiResult: GetResult[DbMaakoodi] =
     GetResult(r =>
